+++ commands/CAdvanced.js
@@ -3,12 +3,50 @@
 const NodeSDK   = require('../common/SDK');
 const Message   = require('../common/Message');
 const Exit      = require('../common/Exit');
+const Helper    = require('../common/Helper');
 const pkg       = require('../package.json');
+
+const CCompany  = require('./CCompany');
+const CUser     = require('./CUser');
 
 class CAdvanced {
 
     constructor(prefs) {
         this._prefs = prefs;
+        this._user = new CUser(this._prefs);
+        this._company = new CCompany(this._prefs);
+    }
+
+    _newco(token, companyName, loginEmail, loginPassword, userFirstname, userLastname, hasAdminRight) {
+
+        let that = this;
+
+        return new Promise((resolve, reject) => {
+
+            that._company._createCompany(token, companyName).then((company) => {
+
+                let options = {
+                    companyId: company.data.id,
+                    isAdmin: hasAdminRight
+                };
+
+                that._user._createSimple(token, loginEmail, loginPassword, userFirstname, userLastname, options).then((user) => {
+
+                    resolve({company: company.data, user: user.data});
+
+                }).catch((errUser) => {
+
+                    that._company._deleteCompany(token, company.data.id, options).then(() => {
+                        reject(errUser);
+                    }).catch((errCompany) => {
+                        reject(errCompany);
+                    });
+                })
+
+            }).catch((err) => {
+                rejec(err);
+            });
+        });
     }
 
     _find(token, id, options) {
@@ -94,6 +132,66 @@ class CAdvanced {
 
                 Message.error(err, options);
                 Exit.error();
+            });
+        }
+        else {
+            Message.notLoggedIn(options);
+            Exit.error();
+        }
+    }
+
+    newco(id, options) {
+        var that = this;
+
+        Message.welcome(options);
+            
+        if(this._prefs.token && this._prefs.user) {
+            Message.loggedin(this._prefs.user, options);
+            Message.action("Wizard for user & company", null, options);
+
+            let companyName = "New Co";
+            let loginEmail = "";
+            let loginPassword = "";
+            let userFirstname = "", userLastname = "";
+            let hasAdminRight = true;
+
+            Message.ask("Enter a company name").then((name) => {
+                companyName = name;
+                return Message.ask("Enter a user email address");
+            }).then((email) => {
+                loginEmail = email;
+                return Message.askPassword("Enter a user password");
+            }).then((password) => {
+                loginPassword = password;
+                return Message.ask("Enter a user first name");
+            }).then((firstname) => {
+                userFirstname = firstname;
+                return Message.ask("Enter a user last name");
+            }).then((lastname) => {
+                userLastname = lastname;
+                return Message.choices("Has user admin right", Helper.YesNo);
+            }).then((isAdmin) => {
+                hasAdminRight = isAdmin;
+
+                let spin = Message.spin(options);
+
+                NodeSDK.start(this._prefs.email, this._prefs.password, this._prefs.host).then(function() {
+                    return that._newco(that._prefs.token, companyName, loginEmail, loginPassword, userFirstname, userLastname, hasAdminRight);
+                }).then(function(json) {
+
+                    Message.unspin(spin);
+
+                    Message.lineFeed();
+                    Message.printSuccess("Company created with Id", json.company.id, options);
+                    Message.printSuccess("User created with Id", json.user.id, options)
+                    Message.success(options);
+
+                }).catch(function(err) {
+                    Message.unspin(spin);
+
+                    Message.error(err, options);
+                    Exit.error();
+                });
             });
         }
         else {